decomp patch: (basename ../../patches/decomp/038-nisplayer-ulong-to-u32.patch .patch)

+++ src/Game/NisPlayer.cpp
@@ -78,7 +78,7 @@ NisPlayer::NisPlayer()
         mAsyncStarted[i] = false;
     }
 
-    unsigned long size = 0;
+    u32 size = 0;
     char* data = (char*)nlLoadEntireFile("art/nis/nis_dict.txt", &size, 0x20, AllocateStart);
     if (data != NULL)
     {
@@ -131,7 +131,7 @@ NisPlayer::NisPlayer()
 
     Reset();
 
-    unsigned long fileSize = 0;
+    u32 fileSize = 0;
     byteCode = nlLoadEntireFile("art/presentation/nis_triggers.byte_code", &fileSize, 0x20, AllocateStart);
     LoadByteCode(byteCode);
     mCamera.m_LetManagerDoUpdate = false;
